pyproject.toml
[tool.poetry]

name = "rainfall"

authors = ["Matt Kucia <maciej@kucia.net>"]
description = "Sample project."
license = "MIT"
readme = "README.md"
repository = "https://github.com/MaciejKucia/rainfall"
version = "0.1"

classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Private :: Do Not Upload",
]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.0"
gunicorn = "^20.1.0"
behave = "^1.2.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3"
isort = "^5.10"
flake8 = "^4.0"

[build-system]
requires = ["poetry>=1,<2"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
Rainfall = "rainfall:main"

[tool.black]
quiet = true

[tool.isort]
profile = "black"
